requirements.txt
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.9.0
pydantic>=2.5.0
numpy>=1.24.0
pandas>=2.0.0
rich>=13.0.0
click>=8.1.0
jinja2>=3.1.0

